fix: Max retention time for usage logs is 30 days...unless since says more

+++ agent/bench.py
@@ -48,8 +48,10 @@ class Bench(Base):
         }
 
     def fetch_sites_info(self, since=None):
+        max_retention_time = (datetime.utcnow() - timedelta(days=30)).timestamp()
+
         if not since:
-            since = (datetime.utcnow() - timedelta(days=30)).timestamp()
+            since = max_retention_time
 
         info = {}
         usage_data = []
@@ -65,7 +67,7 @@ class Bench(Base):
         ]
 
         for file in log_files:
-            if file not in valid_files:
+            if file not in valid_files and os.stat(file).st_mtime > max_retention_time:
                 print(f"Deleting {file}")
                 os.remove(file)
             else:
